move files and fix path
