+++ src/js/events/onLibraryControls.js
